# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateResultDto:
    properties:
      matchEndDate:
        example: "2025-06-01T00:40:00Z"
        type: string
      matchStartDate:
        example: "2025-06-01T00:00:00Z"
        type: string
      opponentScore:
        example: "14"
        type: string
      opponentTeamId:
        example: 4
        type: integer
      score:
        example: "16"
        type: string
    required:
    - matchEndDate
    - matchStartDate
    - opponentTeamId
    type: object
...
  models.ErrorDto:
    properties:
      error:
        example: Error message
        type: string
    type: object
  models.LoginDto:
    properties:
      email:
        example: email@example.com
        type: string
      password:
        example: VeryStrongPassword
        type: string
    required:
    - email
    - password
    type: object
  models.ResultDto:
    properties:
      id:
        example: 7
        type: integer
      matchEndDate:
        example: "2025-06-01T00:40:00Z"
        type: string
      matchStartDate:
        example: "2025-06-01T00:00:00Z"
        type: string
      opponentScore:
        example: "14"
        type: string
      opponentTeamId:
        example: 7
        type: integer
      score:
        example: "16"
        type: string
      seasonId:
        example: 5
        type: integer
      teamId:
        example: 6
        type: integer
      userUuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  models.Role:
    enum:
    - admin
    - moderator
    - user
    type: string
    x-enum-varnames:
    - AdminRole
    - ModeratorRole
    - UserRole
...
  models.SignUpDto:
    properties:
      email:
        example: email@example.com
        type: string
      password:
        example: VeryStrongPassword
        maxLength: 255
        minLength: 8
        type: string
      username:
        example: john_doe_123
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  models.UpdateResultDto:
    properties:
      matchEndDate:
        example: "2025-06-01T00:40:00Z"
        type: string
      matchStartDate:
        example: "2025-06-01T00:00:00Z"
        type: string
      opponentScore:
        example: "14"
        type: string
      score:
        example: "16"
        type: string
    required:
    - matchEndDate
    - matchStartDate
    type: object
...
  models.UserDto:
    properties:
      email:
        example: email@example.com
        type: string
      id:
        example: 526432ea-822b-4b5b-b1b3-34e8ab453e03
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        example: admin
      username:
        example: john_doe_123
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: AurisTFG
    url: https://github.com/AurisTFG
  description: Documentation for MatchMania API
  title: MatchMania API
  version: 0.1.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in
      parameters:
      - description: Log in details
        in: body
        name: loginDto
        required: true
        schema:
          $ref: '#/definitions/models.LoginDto'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      description: Log out
      responses:
        "204":
          description: No Content
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Log out
      tags:
      - auth
  /auth/me:
    get:
      description: Get current user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Refresh token
      responses:
        "204":
          description: No Content
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Refresh token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign up
      parameters:
      - description: Sign up details
        in: body
        name: signUpDto
        required: true
        schema:
          $ref: '#/definitions/models.SignUpDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Sign up
      tags:
      - auth
...
  /seasons/{seasonId}/teams/{teamId}/results:
    get:
      consumes:
      - application/json
      description: Get all results
      parameters:
      - default: "2"
        description: Season ID
        in: path
        name: seasonId
        required: true
        type: string
      - default: "4"
        description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ResultDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Get all results
      tags:
      - results
    post:
      consumes:
      - application/json
      description: Create a result
      parameters:
      - default: "2"
        description: Season ID
        in: path
        name: seasonId
        required: true
        type: string
      - default: "3"
        description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - description: Result object that needs to be created
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/models.CreateResultDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ResultDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Create a result
      tags:
      - results
  /seasons/{seasonId}/teams/{teamId}/results/{resultId}:
    delete:
      consumes:
      - application/json
      description: Delete a result
      parameters:
      - default: "2"
        description: Season ID
        in: path
        name: seasonId
        required: true
        type: string
      - default: "3"
        description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - default: "1"
        description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Delete a result
      tags:
      - results
    get:
      consumes:
      - application/json
      description: Get a result
      parameters:
      - default: "2"
        description: Season ID
        in: path
        name: seasonId
        required: true
        type: string
      - default: "4"
        description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - default: "2"
        description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ResultDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Get a result
      tags:
      - results
    patch:
      consumes:
      - application/json
      description: Update a result
      parameters:
      - default: "2"
        description: Season ID
        in: path
        name: seasonId
        required: true
        type: string
      - default: "4"
        description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - default: "2"
        description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Result object that needs to be updated
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/models.UpdateResultDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResultDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorDto'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorDto'
      summary: Update a result
      tags:
      - results
...
schemes:
- http
swagger: "2.0"